+++ resources/dynamic/Initializer.js
@@ -3,6 +3,7 @@
         instanceId: {%= CurrentADC.InstanceId %},
         headerFixed: {%= CurrentADC.PropValue("headerFixed") %},
         currentQuestion: '{%:= CurrentQuestion.Shortcut %}',
+        maxItemsPerScreen: '{%= CurrentADC.PropValue("maxItemsPerScreen") %}',
         responsiveWidth : '{%= CurrentADC.PropValue("responsiveWidth") %}',
         accordion : {%= CurrentADC.PropValue("accordion") %},
         expandableHeaders : {%= (CurrentADC.PropValue("expandableHeaders") = "1") %},